+++ src/elements/Payment.php
@@ -13,6 +13,8 @@ namespace studioespresso\molliepayments\elements;
 use craft\elements\actions\Restore;
 use craft\helpers\UrlHelper;
 use studioespresso\molliepayments\actions\DeletePaymentAction;
+use studioespresso\molliepayments\actions\ExportAllPaymentsAction;
+use studioespresso\molliepayments\actions\ExportPaymentAction;
 use studioespresso\molliepayments\elements\db\PaymentQuery;
 use studioespresso\molliepayments\MolliePayments;
 
@@ -139,14 +141,21 @@ class Payment extends Element
     {
         return [
             DeletePaymentAction::class,
+            ExportPaymentAction::class,
+            ExportAllPaymentsAction::class
         ];
     }
-    
+
     protected static function defineSortOptions(): array
     {
         return [
             'dateCreated' => \Craft::t('app', 'Date created'),
         ];
+    }
+
+    public static function pluralDisplayName(): string
+    {
+        return Craft::t('mollie-payments', 'Payments');
     }