+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## 1.0.0
+### Added
+- Added action to export selected payments to csv
+- Added action to select all payments to csv
+
+
 ## 1.0.0-beta.2
 ### Added
 - Payment elements can now be deleted from the overview